chore: move all test related stuff under test/
