+++ server/routing_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/Layr-Labs/eigenda-proxy/metrics"
-	"github.com/Layr-Labs/eigenda-proxy/mocks"
+	"github.com/Layr-Labs/eigenda-proxy/test/mocks"
 	"github.com/gorilla/mux"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
