+++ store/generated_key/eigenda/verify/verify_test.go
@@ -32,10 +32,10 @@ func TestCommitmentVerification(t *testing.T) {
 	}
 
 	kzgConfig := kzg.KzgConfig{
-		G1Path:          "../resources/g1.point",
-		G2Path:          "../resources/g2.point",
-		G2TrailingPath:  "../resources/g2.trailing.point",
-		CacheDir:        "../resources/SRSTables",
+		G1Path:          "../../../../resources/g1.point",
+		G2Path:          "../../../../resources/g2.point",
+		G2TrailingPath:  "../../../../resources/g2.trailing.point",
+		CacheDir:        "../../../../resources/SRSTables",
 		SRSOrder:        3000,
 		SRSNumberToLoad: 3000,
 		NumWorker:       uint64(runtime.GOMAXPROCS(0)),
@@ -74,10 +74,10 @@ func TestCommitmentWithTooLargeBlob(t *testing.T) {
 	data := dataRand[:]
 
 	kzgConfig := kzg.KzgConfig{
-		G1Path:          "../resources/g1.point",
-		G2Path:          "../resources/g2.point",
-		G2TrailingPath:  "../resources/g2.trailing.point",
-		CacheDir:        "../resources/SRSTables",
+		G1Path:          "../../../../resources/g1.point",
+		G2Path:          "../../../../resources/g2.point",
+		G2TrailingPath:  "../../../../resources/g2.trailing.point",
+		CacheDir:        "../../../../resources/SRSTables",
 		SRSOrder:        3000,
 		SRSNumberToLoad: 3000,
 		NumWorker:       uint64(runtime.GOMAXPROCS(0)),
