+++ test/mocks/manager.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -package mocks --destination ../mocks/manager.go . IManager
+//	mockgen -package mocks --destination ../test/mocks/manager.go . IManager
 //
 
 // Package mocks is a generated GoMock package.
